+++ src/Phan/Output/Colorizing.php
@@ -118,7 +118,7 @@ class Colorizing
     ) : string {
         $i = 0;
         /** @param array<int,string> $matches */
-        return \preg_replace_callback('/{([A-Z_]+)}|%[sdf]/', static function (array $matches) use ($template, $template_parameters, &$i) : string {
+        return \preg_replace_callback('/(\$?){([A-Z_]+)}|%[sdf]/', static function (array $matches) use ($template, $template_parameters, &$i) : string {
             $j = $i++;
             if ($j >= \count($template_parameters)) {
                 \error_log("Missing argument for colorized output ($template), offset $j");
@@ -133,7 +133,11 @@ class Colorizing
                 // @phan-suppress-next-line PhanPluginPrintfVariableFormatString this is %s, %d, or %f
                 return \sprintf($format_str, $arg);
             }
-            $template = $matches[1];
+            $prefix = $matches[1];
+            $template = $matches[2];
+            if ($prefix) {
+                $arg = $prefix . $arg;
+            }
             return self::colorizeField($template, $arg);
         }, $template);
     }
@@ -159,17 +163,34 @@ class Colorizing
         // @phan-suppress-next-line PhanPluginPrintfVariableFormatString this is %s/%d/%f
         $arg_str = \sprintf($fmt_directive, (string)$arg);
         $color = self::colorForTemplate($template_type);
-        if ($color === null) {
+        if ($color === null || $color === '') {
             \error_log("No color information for template type $template_type");
             return $arg_str;
         }
-        // TODO: Could extend this to support background colors.
-        $color_code = self::STYLES[$color] ?? null;
+        $color_code = self::computeColorCode($color);
         if ($color_code === null) {
             \error_log("Invalid color name ($color) for template type $template_type");
             return $arg_str;
         }
+        // TODO: Could extend this to support background colors.
         return self::colorizeTextWithColorCode($color_code, $arg_str);
+    }
+
+    /**
+     * Compute the color codes (separated by `;`) for the color names.
+     * @param string $color one or more comma separated color names without spaces (e.g. 'none', 'light_gray')
+     */
+    public static function computeColorCode(string $color) : ?string
+    {
+        $color_codes = [];
+        foreach (\explode(',', $color) as $color_component) {
+            $color_code = self::STYLES[$color_component] ?? null;
+            $color_codes[] = $color_code;
+            if ($color_code === null) {
+                return null;
+            }
+        }
+        return \implode(';', $color_codes);
     }
 
     /**
@@ -195,11 +216,41 @@ class Colorizing
     }
 
     /**
+     * @internal
+     */
+    const COLOR_SCHEMES = [
+        'default' => self::class,
+        'eclipse_dark' => \Phan\Output\ColorScheme\EclipseDark::class,
+        'vim' => \Phan\Output\ColorScheme\Vim::class,
+    ];
+
+    /**
+     * @param string $name the name of the color scheme
+     * @return ?array<string,string> maps the template names to their comma separated color codes.
+     */
+    public static function loadColorScheme(string $name) : ?array
+    {
+        if (\array_key_exists($name, self::COLOR_SCHEMES)) {
+            return \constant(self::COLOR_SCHEMES[$name] . '::DEFAULT_COLOR_FOR_TEMPLATE');
+        }
+        return null;
+    }
+
+    /**
      * Initialize the color scheme, merging it with Config::color_scheme
      */
     private static function initColorScheme() : void
     {
         self::$color_scheme = self::DEFAULT_COLOR_FOR_TEMPLATE;
+        $env_color_scheme = \getenv('PHAN_COLOR_SCHEME');
+        if ($env_color_scheme) {
+            $data = self::loadColorScheme($env_color_scheme);
+            if ($data) {
+                self::$color_scheme = $data;
+            } else {
+                \fwrite(\STDERR, "Unknown PHAN_COLOR_SCHEME $env_color_scheme. Supported values: " . \implode(',', \array_keys(self::COLOR_SCHEMES)) . "\n");
+            }
+        }
         foreach (Config::getValue('color_scheme') ?? [] as $template_type => $color_name) {
             if (!\is_string($color_name) || !\array_key_exists($color_name, self::STYLES)) {
                 \error_log("Invalid color name ($color_name)");
@@ -211,5 +262,13 @@ class Colorizing
             }
             self::$color_scheme[$template_type] = $color_name;
         }
+    }
+
+    /**
+     * Used to reset the chosen color scheme in tests.
+     */
+    public static function resetColorScheme() : void
+    {
+        self::$color_scheme = null;
     }
 }